\docType{package}
\name{pkgdownIssue819-package}
\alias{pkgdownIssue819-package}
\title{Demonstrates an issue}
\description{
  Demonstrates an issue.
}
\details{

This is for pkgdown test. The following two phrases should be in separate paragraphs:

\Sexpr[stage=build,results=rd]{paste0(c("First phrase.", "Second phrase."), collapse = "\n\n")}

}
